fix(supplier,image,order): resolve open SonarCloud reliability issues
- Split WorkorderAuthorizationRunner's REQUIRES_NEW/@Transactional methods
  (openRow, applyDecision, parkForReview) into a new
  WorkorderAuthorizationTransactions bean. They were previously called via
  self-invocation (this.openRow(...) etc.), which bypasses the Spring proxy
  and silently disables @Transactional -- so the row-committed-before-vendor-
  call and REQUIRES_NEW-survives-rollback guarantees documented on the class
  never actually held (java:S2229, 4 BLOCKER issues). WorkorderAuthorizationPoller
  now depends on the same bean directly for the calls it was already making
  through the injected proxy correctly.
- MktCatImporter.bodyOrThrow: capture response.body() in a local before the
  null check so the @NonNull return is provably non-null from a single read,
  not two separate calls to the same accessor (java:S2637).
- ImageContentView: override equals/hashCode/toString so the byte[] content
  is compared and reported by value instead of array identity, which the
  default record implementation used (java:S6218).
- ProcurementAvailabilityServiceTest: assert list size before allSatisfy,
  since allSatisfy passes vacuously on an empty list and the test's intent
  requires both lines to be present (java:S5841).
