+++ pos-supplier/src/main/java/com/positivity/supplier/internal/workorderauth/service/WorkorderAuthorizationRunner.java
@@ -11,8 +11,6 @@ import com.positivity.supplier.internal.domain.model.SupplierRequestSpec;
 import com.positivity.supplier.internal.domain.model.SupplierWorkorderAuthorization;
 import com.positivity.supplier.internal.domain.model.WorkorderAuthorizationRequest;
 import com.positivity.supplier.internal.entity.SupplierWorkorderAuthorizationEntity;
-import com.positivity.supplier.internal.enums.WorkorderApprovalStatus;
-import com.positivity.supplier.internal.enums.WorkorderAuthorizationStatus;
 import com.positivity.supplier.internal.exception.SupplierConfigurationException;
 import com.positivity.supplier.internal.registry.AdapterRegistry;
 import com.positivity.supplier.internal.registry.SupplierCodecs;
@@ -21,18 +19,13 @@ import com.positivity.supplier.internal.service.SupplierProfileResolver;
 import com.positivity.supplier.internal.service.SupplierProfileResolver.ResolvedBinding;
 import com.positivity.supplier.internal.spi.SupplierWorkorderAuthorizationPort;
 import com.positivity.supplier.service.model.FleetAuthorizationResponse;
-import java.time.Clock;
-import java.time.Instant;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.jspecify.annotations.NonNull;
 import org.springframework.data.domain.Limit;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Propagation;
-import org.springframework.transaction.annotation.Transactional;
 
 /**
  * Asks a fleet program to authorize work, and records what it said (CAP-323 #1229).
@@ -59,7 +52,6 @@ import org.springframework.transaction.annotation.Transactional;
  * Telling a shop the fleet said no when in fact nobody asked sends it to argue with a fleet manager
  * about a decision that was never made.
  */
-@Slf4j
 @Service
 @RequiredArgsConstructor
 public class WorkorderAuthorizationRunner implements SupplierWorkorderAuthorizationPort {
@@ -77,8 +69,7 @@ public class WorkorderAuthorizationRunner implements SupplierWorkorderAuthorizat
     private final AdapterRegistry adapterRegistry;
     private final SupplierBaseClient baseClient;
     private final SupplierWorkorderAuthorizationRepository authorizationRepository;
-    private final WorkorderAuthorizationPublisher publisher;
-    private final Clock clock;
+    private final WorkorderAuthorizationTransactions transactions;
 
     @Override
     @NonNull
@@ -112,7 +103,10 @@ public class WorkorderAuthorizationRunner implements SupplierWorkorderAuthorizat
                 MichelinS2SWorkorderAuthCodec.class);
         String partnerId = S2SPartnerId.resolve(profileResolver, supplierRef);
 
-        SupplierWorkorderAuthorizationEntity row = openRow(binding, supplierRef, request);
+        // Through the injected bean, NEVER a method on this. A self-invocation would bypass the
+        // Spring proxy and silently disable these methods' @Transactional semantics -- see
+        // WorkorderAuthorizationTransactions for what that cost (SonarCloud java:S2229).
+        SupplierWorkorderAuthorizationEntity row = transactions.openRow(binding, supplierRef, request);
 
         SupplierRequestSpec spec = codec.buildAuthorizationRequest(request, partnerId, API_MAJOR_VERSION);
         SupplierHttpResponse response = baseClient.exchange(SupplierRequests.toHttpRequest(binding, spec));
@@ -121,7 +115,7 @@ public class WorkorderAuthorizationRunner implements SupplierWorkorderAuthorizat
             // Not a denial. The vendor may never have seen the request, or may have created an
             // authorization and failed to tell us -- and those two are indistinguishable from here,
             // which is exactly why a human has to look rather than a loop retrying.
-            return parkForReview(
+            return transactions.parkForReview(
                     row,
                     "vendor exchange failed: " + response.outcome()
                             + (response.failureDetail() == null ? "" : " — " + response.failureDetail()));
@@ -137,136 +131,10 @@ public class WorkorderAuthorizationRunner implements SupplierWorkorderAuthorizat
         } catch (WorkorderAuthDecodeException e) {
             // The vendor answered something. We could not read it. Guessing either way decides
             // whether a shop does work, so neither guess is available.
-            return parkForReview(row, "vendor answer could not be read: " + e.getMessage());
+            return transactions.parkForReview(row, "vendor answer could not be read: " + e.getMessage());
         }
 
-        return applyDecision(row, decision);
-    }
-
-    /**
-     * Records a decision against an existing row and publishes it if terminal.
-     *
-     * <p>Shared with the poller, so an answer that arrives synchronously and one that arrives an
-     * hour later are recorded and published by the same code. Two paths would drift, and the one
-     * that drifts is the asynchronous one, which is also the one nobody watches.
-     */
-    @Transactional
-    @NonNull
-    public SupplierWorkorderAuthorizationEntity applyDecision(
-            @NonNull SupplierWorkorderAuthorizationEntity row, @NonNull SupplierWorkorderAuthorization decision) {
-        Instant now = Instant.now(clock);
-        SupplierWorkorderAuthorizationEntity managed = authorizationRepository
-                .findById(row.getSupplierWorkorderAuthorizationId())
-                .orElse(row);
-
-        managed.setLastPolledAt(now);
-        if (decision.vendorAuthorizationId() != null) {
-            managed.setVendorAuthorizationId(decision.vendorAuthorizationId());
-        }
-
-        switch (decision.status()) {
-            case PENDING -> {
-                managed.setStatus(WorkorderAuthorizationStatus.PENDING);
-                if (decision.pollLocation() != null) {
-                    managed.setPollLocation(decision.pollLocation());
-                }
-            }
-            case GRANTED -> {
-                managed.setStatus(WorkorderAuthorizationStatus.GRANTED);
-                managed.setContractReference(decision.contractReference());
-                managed.setAuthorizedAmount(decision.authorizedAmount());
-                managed.setCurrency(decision.currency());
-                managed.setDecidedAt(now);
-            }
-            case DENIED -> {
-                managed.setStatus(WorkorderAuthorizationStatus.DENIED);
-                managed.setReasonCode(decision.vendorReasonCode());
-                managed.setReasonText(decision.vendorReason());
-                managed.setDecidedAt(now);
-            }
-            case NOT_FOUND -> {
-                managed.setStatus(WorkorderAuthorizationStatus.NOT_FOUND);
-                managed.setReasonCode(decision.vendorReasonCode());
-                managed.setReasonText(decision.vendorReason());
-                managed.setDecidedAt(now);
-            }
-        }
-
-        SupplierWorkorderAuthorizationEntity saved = authorizationRepository.save(managed);
-
-        // Published inside the same transaction as the state change, through the outbox: a decision
-        // recorded but not published would leave the rest of the platform believing the fleet never
-        // answered (ADR-0044 §4).
-        publisher.publishIfTerminal(saved, now);
-        return saved;
-    }
-
-    /**
-     * Parks a row for a human and says why.
-     *
-     * <p>{@code REQUIRES_NEW} so the reason survives whatever the caller does next. A review reason
-     * rolled back with the transaction that discovered it is the one piece of information nobody can
-     * reconstruct afterwards.
-     */
-    @Transactional(propagation = Propagation.REQUIRES_NEW)
-    @NonNull
-    public SupplierWorkorderAuthorizationEntity parkForReview(
-            @NonNull SupplierWorkorderAuthorizationEntity row, @NonNull String reason) {
-        SupplierWorkorderAuthorizationEntity managed = authorizationRepository
-                .findById(row.getSupplierWorkorderAuthorizationId())
-                .orElse(row);
-        managed.setStatus(WorkorderAuthorizationStatus.MANUAL_REVIEW);
-        managed.setReviewReason(truncate(reason));
-        managed.setLastPolledAt(Instant.now(clock));
-        log.warn(
-                "Workorder authorization for workorder {} at {} needs review: {}",
-                managed.getWorkorderId(),
-                managed.getSupplierRef(),
-                reason);
-        return authorizationRepository.save(managed);
-    }
-
-    /**
-     * Opens (or re-opens) the row for this workorder, committed before the vendor is called.
-     *
-     * <p>Re-requesting updates the existing row rather than adding a second. Two rows would let a
-     * denial and a grant coexist for one workorder with nothing to say which one a shop should act
-     * on, and the uniqueness index exists to make that impossible rather than merely unlikely.
-     */
-    @Transactional(propagation = Propagation.REQUIRES_NEW)
-    @NonNull
-    public SupplierWorkorderAuthorizationEntity openRow(
-            @NonNull ResolvedBinding binding,
-            @NonNull SupplierRef supplierRef,
-            @NonNull WorkorderAuthorizationRequest request) {
-        Instant now = Instant.now(clock);
-        UUID vendorProfileId = binding.profile().getVendorProfileId();
-
-        SupplierWorkorderAuthorizationEntity row = authorizationRepository
-                .findByVendorProfileIdAndWorkorderId(vendorProfileId, request.workorderId())
-                .orElseGet(() -> SupplierWorkorderAuthorizationEntity.builder()
-                        .vendorProfileId(vendorProfileId)
-                        .workorderId(request.workorderId())
-                        .approvalStatus(WorkorderApprovalStatus.NOT_REQUESTED)
-                        .approvalAttempts(0)
-                        .build());
-
-        row.setSupplierRef(supplierRef.value());
-        row.setStatus(WorkorderAuthorizationStatus.PENDING);
-        row.setRequestedAt(now);
-        // Cleared on re-request: a stale reason from the previous attempt read as if it described
-        // this one.
-        row.setReviewReason(null);
-        row.setReasonCode(null);
-        row.setReasonText(null);
-        row.setDecidedAt(null);
-        row.setPollLocation(null);
-        return authorizationRepository.save(row);
-    }
-    /** Keeps a vendor's error text inside the column it is stored in. */
-    @NonNull
-    private static String truncate(@NonNull String reason) {
-        return reason.length() <= 1000 ? reason : reason.substring(0, 1000);
+        return transactions.applyDecision(row, decision);
     }
 
     /**
